requirements.txt
opencv-python==4.9.0.80
Pillow==11.2.1
face_recognition==1.3.0
